requirement.txt
# pip install -r requirements.txt
Cython
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.2
Pillow
torch>=1.6.0
torchvision>=0.7.0
pycocotools>=2.0  # COCO mAP